% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRegTreeData.R
\name{extractRegTreeData}
\alias{extractRegTreeData}
\title{Extract regeneration data from IFN3/4}
\usage{
extractRegTreeData(regData, height.cm = FALSE, subsetVars = TRUE)
}
\arguments{
\item{regData}{A data frame returned by \code{readRegenera}}

\item{height.cm}{Flag to return height in 'cm' instead of 'meters'}

\item{subsetVars}{Flag to filter data columns}
}
\description{
Extract regeneration data from IFN3/4
}
