% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDatosMap.R
\name{readDatosMap}
\alias{readDatosMap}
\alias{readDatosMapIFN3}
\title{Reads map data (DatosMap)}
\usage{
readDatosMapIFN3(prov, DBFdir = "DBF", plotTypeIFN3 = c("A1", "NN"))
}
\arguments{
\item{prov}{Character vector with codes of provinces (e.g. \code{c("01","03")}) or \code{"all"} for all provinces.}

\item{DBFdir}{String with the path where provinces are stored}

\item{plotTypeIFN3}{Subset of plot types to include (see \code{\link{readPiesMayoresIFN3}})}
}
\value{
a data frame with map data
}
\description{
Reads map data (DatosMap)
}
