% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIFN2.R
\name{readIFN2}
\alias{readIFN2}
\alias{readPiesMenoresIFN2}
\alias{readMatorralIFN2}
\alias{readPiesMayoresIFN2}
\alias{readPCParcelaIFN2}
\title{Functions to read IFN2 data}
\usage{
readPiesMenoresIFN2(prov, DBFdir = "DBF", subsetVars = TRUE)

readMatorralIFN2(prov, DBFdir = "DBF", height.cm = FALSE, subsetVars = TRUE)

readPiesMayoresIFN2(
  prov,
  DBFdir = "DBF",
  plotIDs = NULL,
  rem.nodensity = FALSE,
  rem.nospecies = FALSE,
  rem.noheight = FALSE,
  height.cm = FALSE,
  subsetVars = TRUE
)

readPCParcelaIFN2(prov, DBFdir = "DBF", plotIDs = NULL)
}
\arguments{
\item{prov}{Character vector with codes of provinces (e.g. \code{c("01","03")}).}

\item{DBFdir}{String with the path where provinces are stored}

\item{subsetVars}{Flag to filter data columns}

\item{height.cm}{Flag to return height in 'cm' instead of 'meters'}

\item{plotIDs}{String vector with a selection of plots}

\item{rem.nodensity}{Flag to remove records without density}

\item{rem.nospecies}{Flag to remove records without species identity}

\item{rem.noheight}{Flag to remove records without tree height}
}
\value{
a data frame
}
\description{
Functions to read IFN2 data
}
