% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMatorral.R
\name{readMatorral}
\alias{readMatorral}
\alias{readMatorralIFN2}
\title{Read shrub data}
\usage{
readMatorral(accessFiles, provincias = NULL, plotType = c("A1", "NN"),
  plotIDs = NULL, rem.nospecies = FALSE, height.cm = TRUE,
  subsetVars = TRUE)

readMatorralIFN2(prov, DBFdir = "DBF", height.cm = TRUE,
  subsetVars = TRUE)
}
\arguments{
\item{accessFiles}{A character vector of access files to be read.}

\item{provincias}{Numeric or character vector with the province numbers corresponding to \code{accessFiles}}

\item{plotType}{Subset of plot types to include:}

\item{plotIDs}{String vector with a selection of plots}

\item{rem.nospecies}{Flag to remove records without species identity}

\item{height.cm}{Flag to return height in 'cm' instead of 'meters'}

\item{subsetVars}{Flag to filter data columns}
}
\value{
a data frame with shrub data
}
\description{
Reads shrub data from MSAccess database
}
