% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPCParcela.R
\name{readPCParcela}
\alias{readPCParcela}
\alias{readPCParcelaIFN2}
\alias{readPCParcelaIFN3}
\alias{readPCParcelaIFN4}
\title{Reads plot data (PC parcelas)}
\usage{
readPCParcelaIFN2(prov, DBFdir = "DBF", plotIDs = NULL)

readPCParcelaIFN3(
  prov,
  DBFdir = "DBF",
  plotTypeIFN3 = c("A1", "NN"),
  plotIDs = NULL
)

readPCParcelaIFN4(accessFiles, plotTypeIFN4 = c("A1", "N"), plotIDs = NULL)
}
\arguments{
\item{prov}{Character vector with codes of provinces (e.g. \code{c("01","03")}) or \code{"all"} for all provinces.}

\item{DBFdir}{String with the path where provinces are stored}

\item{plotIDs}{String vector with a selection of plots}

\item{plotTypeIFN3}{Subset of plot types to include (see \code{\link{readPiesMayoresIFN3}})}

\item{accessFiles}{A character vector of access files to be read.}

\item{plotTypeIFN4}{Subset of plot types to include (see \code{\link{readPiesMayoresIFN3}})}
}
\value{
a data frame with plot data
}
\description{
Reads plot data (PC parcelas)
}
