% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPiesMayores.R
\name{readPiesMayores}
\alias{readPiesMayores}
\alias{readPiesMayoresIFN2}
\alias{readPiesMayoresIFN3}
\alias{readPiesMayoresIFN4}
\title{Reads the table of pies mayores}
\usage{
readPiesMayoresIFN2(prov, DBFdir = "DBF", plotIDs = NULL,
  rem.nodensity = FALSE, rem.nospecies = FALSE, rem.noheight = FALSE,
  height.cm = FALSE, subsetVars = TRUE)

readPiesMayoresIFN3(accessFiles, plotTypeIFN3 = c("A1", "NN"),
  plotIDs = NULL, include.IFN2 = FALSE, rem.nospecies = FALSE,
  rem.noheight = FALSE, rem.nodensity = FALSE,
  rem.codes32 = character(0), rem.codes33 = character(0),
  height.cm = FALSE, subsetVars = TRUE)

readPiesMayoresIFN4(accessFiles, plotTypeIFN4 = c("A1", "NN"),
  plotIDs = NULL, include.IFN3 = FALSE, rem.nospecies = FALSE,
  rem.noheight = FALSE, rem.nodensity = FALSE,
  rem.codes43 = character(0), rem.codes44 = character(0),
  height.cm = FALSE, subsetVars = TRUE, vars.IFN3 = c("OrdenIf3",
  "Especie", "Dn1", "Dn2", "Ht", "Forma"))
}
\arguments{
\item{prov}{Character vector with codes of provinces (e.g. \code{c("01","03")}) or \code{"all"} for all provinces.}

\item{DBFdir}{String with the path where provinces are stored}

\item{plotIDs}{String vector with a selection of plots}

\item{rem.nodensity}{Flag to remove records without density}

\item{rem.nospecies}{Flag to remove records without species identity}

\item{rem.noheight}{Flag to remove records without tree height}

\item{height.cm}{Flag to return height in 'cm' instead of 'meters'}

\item{subsetVars}{Flag to filter data columns}

\item{accessFiles}{A character vector of access files to be read.}

\item{plotTypeIFN3}{Subset of plot types to include:
\itemize{
\item{N - Se levantan por primera vez en el IFN-3 y son utilizadas para el calculo de existencias actuales.}
\item{A - Parcelas levantadas en el IFN-2 que volveran a ser apeadas en el IFN-3. Existen cinco casos:
  \itemize{
    \item{
       A1 . Parcelas en las que se localice el rejon y esta bien implantado. Se utilizan para la comparacion de inventarios y para el calculo de existencias actuales.
    }
    \item{
       A3 . Parcela apeada en IFN2, pinchazo IFN2 y coordenada UTM situadas en diferentes teselas. Se cumplimentan dos estadillos:
       \itemize{
         \item{
           A3C . Parcela para la Comparacion de Inventarios. Se levantan en el punto donde se encuentra el rejon del IFN-2.
         }
         \item{
           A3E. Parcelas para el calculo de Existencias Actuales. Se levantan en el emplazamiento del IFN-3 ( Coordenada UTM)
         }
       }
    }
    \item{
      A4 . Parcelas en las que no se localizo el rejon. Se utilizan para el calculo de existencias actuales.
    }
    \item{
      A4C . Parcelas en las que no se localice el rej?n por cambio totales en la cubierta forestal de la tesela que contiene la parcela (p?rdida total de la masa
muestreada en el IFN2 o incorporaci?n de una nueva masa forestal inexistente en el IFN2). Se utilizar?n para la comparaci?n de inventarios y para el c?lculo de existencias actuales.
    }
    \item{
      A6C . Parcelas en las que no se localice el rej?n por cambios totales en la
cubierta forestal de la tesela que contiene la parcela (p?rdida total de la masa
muestreada en el IFN2 e incorporaci?n de una nueva masa forestal). Se utilizar?n para la comparaci?n de inventarios y para el c?lculo de existencias actuales.
    }
  }
}
}}

\item{include.IFN2}{Flag to include diameter and height measured in IFN2}

\item{rem.codes32}{set of tree codes in IFN2 corresponding to records to be removed:
\itemize{
  \item{'0' or '000' - Tree not tallied in IFN2 because of its small size.}
  \item{'999' - Tree that should have been tallied in IFN2 but was not.}
}}

\item{rem.codes33}{set of tree codes in IFN3 corresponding to records to be removed
\itemize{
  \item{'000' - Tree tallied in IFN2 but cut between the two inventories.}
  \item{'444' - Tree tallied in IFN2 and the stump is not found or it is in advanced decomposition state.}
  \item{'777' - Cork oak tree tallied in IFN2 and scorched between the two inventories.}
  \item{'888' - Tree tallied in IFN2 and dead between the two inventories.}
  \item{'999' - Tree tallied in IFN2 but should have not been included.}
}}

\item{plotTypeIFN4}{Subset of plot types to include:}

\item{include.IFN3}{Flag to include diameter and height measured in IFN3}

\item{rem.codes43}{set of tree codes in IFN3 corresponding to records to be removed:
\itemize{
  \item{'0' or '000' - Tree not tallied in IFN3 because of its small size.}
  \item{'999' - Tree that should have been tallied in IFN3 but was not for some reason.}
}}

\item{rem.codes44}{set of tree codes in IFN4 corresponding to records to be removed
\itemize{
  \item{'000' - Tree tallied in IFN3 but cut between the two inventories (stump is found).}
  \item{'444' - Tree tallied in IFN3 and the stump is not found or it is in advanced decomposition state.}
  \item{'777' - Cork oak tree tallied in IFN3 and scorched between the two inventories.}
  \item{'888' - Tree tallied in IFN3 and dead between the two inventories.}
  \item{'999' - Tree tallied in IFN3 but should have not been included.}
}}

\item{vars.IFN3}{String vector with column names of variables in the IFN3 table}
}
\value{
a data frame
}
\description{
Reads the table of pies mayores
}
\examples{

\dontrun{
accessfilesIFN3 = list.files("//SERVERPROCESS/Miquel/Datasets/IFN/IFN3/MSAccess/",
                             pattern = "*[.]accdb", full.names = T)

treeData = readPiesMayoresIFN3(accessfilesIFN3[prov])

}
}
