% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRegenera.R
\name{readRegenera}
\alias{readRegenera}
\alias{extractRegTreeData}
\title{Read regeneration data}
\usage{
readRegenera(
  accessFiles,
  provincias = NULL,
  plotType = c("A1", "NN"),
  plotIDs = NULL,
  rem.nospecies = FALSE,
  subsetVars = TRUE
)

extractRegTreeData(regData, height.cm = FALSE, subsetVars = TRUE)
}
\arguments{
\item{accessFiles}{A character vector of access files to be read.}

\item{provincias}{Numeric or character vector with the province numbers corresponding to \code{accessFiles}}

\item{plotType}{Subset of plot types to include:}

\item{plotIDs}{String vector with a selection of plots}

\item{rem.nospecies}{Flag to remove records without species identity}

\item{subsetVars}{Flag to filter data columns}

\item{regData}{A data frame returned by \code{readRegenera}}

\item{height.cm}{Flag to return height in 'cm' instead of 'meters'}
}
\value{
a data frame with regeneration data
}
\description{
Reads regeneration data from MSAccess database (IFN3-IFN4)
}
