% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIFN2.R
\name{readIFN2}
\alias{readIFN2}
\alias{readPiesMenoresIFN2}
\alias{readMatorralIFN2}
\alias{readPiesMayoresIFN2}
\alias{readPCParcelaIFN2}
\title{Functions to read IFN2 data}
\usage{
readPiesMenoresIFN2(prov, DBFdir = "DBF", renameVars = TRUE)

readMatorralIFN2(prov, DBFdir = "DBF", renameVars = TRUE)

readPiesMayoresIFN2(prov, DBFdir = "DBF", renameVars = TRUE)

readPCParcelaIFN2(prov, DBFdir = "DBF", renameVars = TRUE)
}
\arguments{
\item{prov}{Character vector with codes of provinces (e.g. \code{c("01","03")}).}

\item{DBFdir}{String with the path where provinces are stored}

\item{renameVars}{Flag to rename data columns to make them similar to IFN3/IFN4}
}
\value{
a data frame
}
\description{
Functions to read IFN2 data
}
